requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appdirs=1.4.4=pyhd3eb1b0_0
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py311hb9e55a9_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py311h6c40b1e_1002
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.05.30=hecd8cb5_0
certifi=2023.5.7=py311hecd8cb5_0
cffi=1.15.1=py311h6c40b1e_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.3=pyhd8ed1ab_0
contourpy=1.0.5=py311ha357a0b_0
cryptography=39.0.1=py311hf6deb26_2
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py311h814d153_0
decorator=5.1.1=pyhd8ed1ab_0
dgl=1.1.0=py311_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.3=h0a44026_0
filelock=3.9.0=py311hecd8cb5_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
gettext=0.21.0=he85b6c0_1
giflib=5.2.1=h6c40b1e_3
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py311h1c2e9e1_0
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
idna=3.4=py311hecd8cb5_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
intel-openmp=2023.1.0=ha357a0b_43547
ipykernel=6.23.1=pyh736e0ef_0
ipython=8.14.0=pyhd1c38e8_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py311hecd8cb5_0
joblib=1.2.0=py311hecd8cb5_0
jpeg=9e=h6c40b1e_1
jupyter_client=8.2.0=pyhd8ed1ab_0
jupyter_core=4.12.0=py311h6eed73b_0
kiwisolver=1.4.4=py311hcec6c5f_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_0
libffi=3.4.4=hecd8cb5_0
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=hca72f7f_2
libidn2=2.3.4=h6c40b1e_0
libpng=1.6.39=h6c40b1e_0
libsodium=1.0.18=hbcb3906_1
libtasn1=4.19.0=h6c40b1e_0
libtiff=4.5.0=hcec6c5f_2
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.4=hf6ce154_1
libwebp-base=1.2.4=h6c40b1e_1
libxml2=2.10.3=h930c0e2_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
markupsafe=2.1.1=py311h6c40b1e_0
matplotlib=3.7.1=py311hecd8cb5_1
matplotlib-base=3.7.1=py311h11e8b89_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h59209a4_43558
mkl-service=2.4.0=py311h6c40b1e_1
mkl_fft=1.3.6=py311hdb55bb0_1
mkl_random=1.2.2=py311hdb55bb0_1
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.2.1=py311hecd8cb5_0
munkres=1.1.4=py_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=h230ac6f_1
networkx=2.8.4=py311hecd8cb5_1
numexpr=2.8.4=py311h728a8a3_1
numpy=1.24.3=py311h728a8a3_1
numpy-base=1.24.3=py311h53bf9ac_1
openh264=2.1.1=h8346a28_0
openssl=3.0.8=hca72f7f_0
packaging=23.0=py311hecd8cb5_0
pandas=1.5.3=py311hc5848a5_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py311hcec6c5f_0
pip=23.1.2=py311hecd8cb5_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
psutil=5.9.0=py311h6c40b1e_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pyhd8ed1ab_0
pyopenssl=23.0.0=py311hecd8cb5_0
pyparsing=3.0.9=py311hecd8cb5_0
pysocks=1.7.1=py311hecd8cb5_0
python=3.11.3=hf27a42d_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.11=2_cp311
pytorch=2.0.1=py3.11_0
pytz=2022.7=py311hecd8cb5_0
pyzmq=25.1.0=py311hcec6c5f_0
readline=8.2=hca72f7f_0
requests=2.29.0=py311hecd8cb5_0
scikit-learn=1.2.2=py311hcec6c5f_0
scipy=1.10.1=py311h224febf_1
setuptools=67.8.0=py311hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py311hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h5d9f67b_0
torchaudio=2.0.2=py311_cpu
torchvision=0.15.2=py311_cpu
tornado=6.3.2=py311h2725bcf_0
tqdm=4.65.0=py311h85bffb1_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.6.3=py311hecd8cb5_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311hecd8cb5_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py311hecd8cb5_0
xz=5.4.2=h6c40b1e_0
zeromq=4.3.4=he49afe7_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0